requirements.txt
aniso8601==4.0.1
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.6
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
pathlib2==2.3.3
pkg-resources==0.0.0
pluggy==0.8.0
psycopg2-binary==2.7.6.1
py==1.7.0
PyJWT==1.7.0
pytest==4.0.1
pytz==2018.7
six==1.11.0
Werkzeug==0.14.1
